# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": ["code", "message", "details"],
        "title": "ApiError",
        "type": "object"
      },
      "BranchRequest": {
        "additionalProperties": false,
        "properties": {
          "branch": {
            "maxLength": 72,
            "minLength": 1,
            "title": "Branch",
            "type": "string"
          },
          "create": {
            "default": false,
            "title": "Create",
            "type": "boolean"
          }
        },
        "required": ["branch"],
        "title": "BranchRequest",
        "type": "object"
      },
...
      "EditableResourceModel": {
        "additionalProperties": false,
        "properties": {
          "acceptance_criterion_es": {
            "title": "Acceptance Criterion Es",
            "type": "string"
          },
          "alternative_labels_es": {
            "items": {
              "type": "string"
            },
            "title": "Alternative Labels Es",
            "type": "array"
          },
          "author": {
            "title": "Author",
            "type": "string"
          },
          "class_iri": {
            "title": "Class Iri",
            "type": "string"
          },
          "definition_es": {
            "title": "Definition Es",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "evidence": {
            "title": "Evidence",
            "type": "string"
          },
          "form_values": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Form Values",
            "type": "object"
          },
          "iri": {
            "title": "Iri",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "module_id": {
            "title": "Module Id",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "preferred_label_es": {
            "title": "Preferred Label Es",
            "type": "string"
          },
          "question_text_es": {
            "title": "Question Text Es",
            "type": "string"
          },
          "range": {
            "title": "Range",
            "type": "string"
          },
          "reading_direction_es": {
            "title": "Reading Direction Es",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "valid_example": {
            "title": "Valid Example",
            "type": "string"
          }
        },
        "required": [
          "iri",
          "kind",
          "preferred_label_es",
          "alternative_labels_es",
          "definition_es",
          "module_id",
          "status",
          "evidence",
          "author",
          "reading_direction_es",
          "valid_example",
          "domain",
          "range",
          "class_iri",
          "source_id",
          "question_text_es",
          "acceptance_criterion_es",
          "form_values",
          "path"
        ],
        "title": "EditableResourceModel",
        "type": "object"
      },
...
      "PullRequestRequest": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "draft": {
            "default": true,
            "title": "Draft",
            "type": "boolean"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": ["title", "body"],
        "title": "PullRequestRequest",
        "type": "object"
      },
...
      "SemanticDiffModel": {
        "additionalProperties": false,
        "properties": {
          "added_quads": {
            "items": {
              "$ref": "#/components/schemas/SemanticQuadModel"
            },
            "title": "Added Quads",
            "type": "array"
          },
          "added_resources": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Added Resources",
            "type": "array"
          },
          "affected_modules": {
            "items": {
              "type": "string"
            },
            "title": "Affected Modules",
            "type": "array"
          },
          "base": {
            "title": "Base",
            "type": "string"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/ResourceChangeModel"
            },
            "title": "Changes",
            "type": "array"
          },
          "deprecated_resources": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Deprecated Resources",
            "type": "array"
          },
          "head": {
            "title": "Head",
            "type": "string"
          },
          "modified_resources": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Modified Resources",
            "type": "array"
          },
          "potentially_impacted_questions": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Potentially Impacted Questions",
            "type": "array"
          },
          "removed_quads": {
            "items": {
              "$ref": "#/components/schemas/SemanticQuadModel"
            },
            "title": "Removed Quads",
            "type": "array"
          }
        },
        "required": [
          "base",
          "head",
          "added_resources",
          "modified_resources",
          "deprecated_resources",
          "added_quads",
          "removed_quads",
          "changes",
          "affected_modules",
          "potentially_impacted_questions"
        ],
        "title": "SemanticDiffModel",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Stable API over the canonical RDF/Git workspace, including controlled proposal authoring, semantic review and the synchronized canonical agent contract.",
    "title": "Enterprise Ontology Workbench API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/authoring/state": {
      "get": {
        "operationId": "authoring_state_api_authoring_state_get",
        "parameters": [
          {
            "in": "query",
            "name": "iri",
            "required": true,
            "schema": {
              "title": "Iri",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditableResourceModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Authoring State",
        "tags": ["editing"]
      }
    },
...
    "/api/diff": {
      "get": {
        "operationId": "semantic_diff_api_diff_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemanticDiffModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Semantic Diff",
        "tags": ["editing"]
      }
    },
    "/api/git/branch": {
      "post": {
        "operationId": "git_branch_api_git_branch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitStatusModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Git Branch",
        "tags": ["editing"]
      }
    },
    "/api/git/commit": {
      "post": {
        "operationId": "git_commit_api_git_commit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Git Commit",
        "tags": ["editing"]
      }
    },
    "/api/git/pull_request": {
      "post": {
        "operationId": "git_pull_request_api_git_pull_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PullRequestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PullRequestResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Git Pull Request",
        "tags": ["editing"]
      }
    },
    "/api/git/status": {
      "get": {
        "operationId": "git_status_api_git_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitStatusModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Git Status",
        "tags": ["editing"]
      }
    },
...
    "/api/relations": {
      "post": {
        "operationId": "write_relation_api_relations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Write Relation",
        "tags": ["editing"]
      }
    },
    "/api/relations/draft": {
      "delete": {
        "operationId": "delete_relation_api_relations_draft_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Delete Relation",
        "tags": ["editing"]
      }
    },
    "/api/resources": {
      "patch": {
        "operationId": "write_term_api_resources_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TermWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Write Term",
        "tags": ["editing"]
      },
      "post": {
        "operationId": "write_term_api_resources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TermWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Write Term",
        "tags": ["editing"]
      }
    },
    "/api/resources/deprecate": {
      "post": {
        "operationId": "deprecate_resource_api_resources_deprecate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeprecationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Deprecate Resource",
        "tags": ["editing"]
      }
    },
...
    "/api/resources/individual": {
      "post": {
        "operationId": "write_individual_api_resources_individual_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndividualWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Write Individual",
        "tags": ["editing"]
      }
    },
...
    "/api/review": {
      "get": {
        "operationId": "proposal_review_api_review_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalReviewModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Proposal Review",
        "tags": ["editing"]
      }
    },
...
  }
}